Introduce Mikoto.
